before uint32_t 4byte ix

+++ Core/Libraries/Source/WWVegas/WWLib/bittype.h
@@ -39,12 +39,16 @@
 
 typedef unsigned char	uint8;
 typedef unsigned short	uint16;
-typedef unsigned long	uint32;
+// Phase D8 — was `unsigned long`, which is 8 bytes on 64-bit macOS/Linux and
+// silently corrupts every binary file format that lays out u32 fields
+// (ChunkHeader, W3D struct headers, etc.). Use the fixed-width type so the
+// file layout matches the on-disk one across platforms.
+typedef unsigned int	uint32;
 typedef unsigned int    uint;
 
 typedef signed char		sint8;
 typedef signed short		sint16;
-typedef signed long		sint32;
+typedef signed int		sint32;
 typedef signed int      sint;
 
 typedef float				float32;
